+++ packages/dumb/src/components/stats/insulin/insulin-stat.tsx
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022-2023, Diabeloop
+ * Copyright (c) 2022-2024, Diabeloop
  *
  * All rights reserved.
  *
@@ -104,6 +104,7 @@ const InsulinStat: FunctionComponent<TotalInsulinStatProps> = (props) => {
           display="flex"
           justifyContent="space-between"
           alignItems={isDisabledTotalInsuline}
+          className={styles.valueColor}
         >
           <span className={styles.titleTotal}>
             {totalInsulin}
@@ -129,16 +130,17 @@ const InsulinStat: FunctionComponent<TotalInsulinStatProps> = (props) => {
                 display="flex"
                 justifyContent="space-between"
                 alignItems="center"
+                className={styles.valueColor}
               >
                 <Chip
                   label={`${entry.value > 0 ? entry.valueString : '0'} ${entry.units}`}
                   variant="outlined"
-                  className={`${styles[`rowsTotalInsulin-${entry.id}`]}`}
+                  className={styles.valueColor}
                   size="small"
                   sx={{ marginRight: theme.spacing(2) }}
                 />
                 <Box
-                  className={`${styles.rowPercent} ${styles[`rowsTotalInsulin-${entry.id}`]}`}
+                  className={`${styles.rowPercent}`}
                   width="50px"
                   alignItems={entry.value === 0 ? 'center' : 'baseline'}
                 >
@@ -167,6 +169,7 @@ const InsulinStat: FunctionComponent<TotalInsulinStatProps> = (props) => {
           <Box
             display="flex"
             alignItems={isDisabledWeight ? 'center' : 'baseline'}
+            className={styles.valueColor}
           >
             <span className={getOutputValueClasses()}>
               {weight}
@@ -186,6 +189,7 @@ const InsulinStat: FunctionComponent<TotalInsulinStatProps> = (props) => {
           <Box
             display="flex"
             alignItems={isDisabledWeight ? 'center' : 'baseline'}
+            className={styles.valueColor}
           >
             <span className={getOutputValueClasses()}>
               {getDailyDosePerWeight()}