+++ packages/tideline/js/plot/reservoir.js
@@ -23,6 +23,7 @@ import danaPumpIcon from 'dana-pump.svg'
 import insightPumpIcon from 'insight-pump.svg'
 import kaleidoPumpIcon from 'kaleido-pump.svg'
 import medisafePumpIcon from 'medisafe-pump.svg'
+import { DEFAULT_IMAGE_MARGIN, DEFAULT_OPTIONS_SIZE } from './util/eventsConstants'
 
 const getReservoirChangeIcon = (pumpManufacturer) => {
   const manufacturerUpperCase = pumpManufacturer.toUpperCase()
@@ -41,10 +42,12 @@ const getReservoirChangeIcon = (pumpManufacturer) => {
 
 function plotReservoirChange(pool, opts) {
   const d3 = window.d3
-  const height = pool.height() / 5
+  const height = pool.height() - DEFAULT_IMAGE_MARGIN
   const width = 40
 
   const xPos = (d) => opts.xScale(d.epoch)
+
+  opts.size = opts.size ?? DEFAULT_OPTIONS_SIZE
 
   function reservoir(selection) {
     opts.xScale = pool.xScale().copy()
@@ -76,7 +79,7 @@ function plotReservoirChange(pool, opts) {
         .append('image')
         .attr({
           'x': (d) => xPos(d) - (width / 2) ,
-          'y': 0,
+          'y': pool.height() / 2 - opts.size / 2,
           width,
           height,
           'xlink:href': (reservoirChange) => getReservoirChangeIcon(reservoirChange.pump.manufacturer)