+++ CHANGELOG.md
@@ -5,6 +5,7 @@ It is based on Tidepool Blip 1.27.
 ## 3.4.0-beta - 2024-04-05
 ### Changed
 - YLP-2866 Display events in a dedicated section in Daily view
+- YLP-2974 Update colors in Daily and Trends views to match DBL theme
 
 ### Engineering use
 - YLP-2948 Updated failing tests because of the time change