+++ CHANGELOG.md
@@ -6,7 +6,9 @@ It is based on Tidepool Blip 1.27.
 ### Changed
 - YLP-2866 Display events in a dedicated section in Daily view
 - YLP-2974 Update colors in Daily and Trends views to match DBL theme
+- YLP-2992 As a user, I want to see the daily numbers of rescue carbs intake / manual boluses and not the average
 - YLP-3019 New design for stats
+
 
 ### Engineering use
 - YLP-2948 Updated failing tests because of the time change